requirements.txt
altgraph==0.17.4
black==24.10.0
click==8.1.7
colorama==0.4.6
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pywin32-ctypes==0.2.3
setuptools==75.3.0
tk==0.1.0
